style: rename extended_fetcher mod to eigenda_fetcher
